Apply change requests
